Change file size return type to int32_t in FTP implementation and update logging format

+++ lib/FileSystem/fnFsFTP.cpp
@@ -187,8 +187,8 @@ FileHandler *FileSystemFTP::cache_file(const char *path, const char *mode)
         return nullptr;
 
     // Get file size from FTP server
-    size_t filesize = _ftp->get_file_size(path);
-    Debug_printf("File size reported by server: %u bytes\n", filesize);
+    int32_t filesize = _ftp->get_file_size(path);
+    Debug_printf("File size reported by server: %lu bytes\n", filesize);
 
     // Open FTP file
     Debug_println("Initiating file RETR");
